raw-transform/requirements.txt
google-cloud-bigquery
flask
functions-framework==3.*
google-cloud-secret-manager
google-cloud-bigquery